The real checkpoint4 part1 (#29)
* Just added the things needed

To do the checkpoint 4  test!

* Just got the first half the other half left!

* adding the implementation for ComputeApi and Prototype

* Make the compute Api work

I added the required implementation and made sure all of them are being used and passed the junit test!

* removed them for now!

Will be added after implemented all three apis and checkpoints passed!

* Now, it should fix the error!

For some reason, it was saying it can't find the conceptualApi not imported in my conceptual api but now it's fixed!

* FORGOT THE INDENTATION!