The real checkpoint4 final part 4 (#39)
* Fixed the problem of framwork!

Fixed the infinite loop!

* Make the correction of the method that aren't being used!

Make sure you fix up the feedback from checkpoint 3 (removing methods from the APIs that aren't called from the prototypes, plus any other feedback).

* BIG FIXES!

Fix the mixup of incorrect linking with api's and unnessary comments clogging up the code base time!

* trying again?

* trying again part 2.

For some reason, more errors shows up in checkpoint test 2 and 3 due to not found api's like conceputal and others but are clearly there???

* moved it down to pass the checkstyle

* fix the problem

added the correct parameter for one of the apis and testdatastoreapi correctly as well and passes both checkpoint 2 and 3

* Fixed the checkpoint 4 from not passing to passing!

The reason why it was failing due to I was fixing checkpoint 5. Now it's corrected and wait again for results to come out!