Checkpoint5 part 3 (#40)
* On hold for rn

* Now it will be fixed before add the correct code!

* Pass the test

IT gave me red printout but its working it's just system.out.error instead of print and the throw exception is working correctly as well as passing the junit!

* fix the problem

* Got it to pass correctly!

It doing what it is intended to be doing.

* will fix it soon!

* fixed up the merge conflict and error handling

* passed checkpoint4test

* FIXED that problem!

I created an interface call coordinator and a class called coordinatorApi to make checkpoint 4 pass and checkpoint 5 to pass as well! With the corrected output now!