Cap stone only (Checkpoint 6 feedback) (#50)
* will fix later

* will fix later

* added updates to handle big numbers!

* fix checkstyle in captilzing TestingExample from testingExample

* Fix another checkstyle

* Create settings.json

* Fix the feedback from checkpoint 5! AGain!

* Redid capstone part 1/2 of it!

Where it can handle: The engine should accept arbitrarily large inputs

* To be able to handle

Can handle: The engine should accept arbitrarily large inputs

* Fix the checkstyle issue!

* I forgot to change the value of the larger number

The larger number was 123456789012345678901234567890 and it was able to handle it and result was an expected! 23264

* I fixed the feedback for checkpoint 6.

According to the professors feedback: I made it where it passes the feedback the professor gave to me: Make sure your unit tests are still unit tests; each component should have at least one test where it's the only real component under test, with mocks used for everything else
Make sure all methods on each API are actually used by clients of the API; if you need to update your prototypes to reflect the updated workflows, that's fine (in other words, any method that isn't called by either non-prototype code in src/ or the ManualTestingFramework should be removed).

* Fixed the checkstyle error

Removed it where I was using: import static org.mockito.Mockito.* (Which was wrong due to incorrect import. When it was supposed to so what exact import I am using for mockito.